fixed int for size in lca stuff

+++ library/graphs/lca_jump.cpp
@@ -21,7 +21,7 @@ struct LCAJump {
         dfs(root);
     }
     void dfs(int src = 0) {
-        for (int i = 1; i < par.size(); i++) {
+        for (int i = 1; i < (int) par.size(); i++) {
             par[i][src] = par[i - 1][par[i - 1][src]];
         }
         for (int nxt: adj[src]) {
@@ -31,7 +31,7 @@ struct LCAJump {
         }
     }
     int jump(int x, int d) {
-        for (int i = 0; i < par.size(); i++) {
+        for (int i = 0; i < (int) par.size(); i++) {
             if ((d >> i) & 1) {
                 x = par[i][x];
             }
@@ -42,7 +42,7 @@ struct LCAJump {
         if (depth[x] < depth[y]) std::swap(x, y);
         x = jump(x, depth[x] - depth[y]);
         if (x == y) return x;
-        for (int i = par.size() - 1; i >= 0; i--) {
+        for (int i = (int) par.size() - 1; i >= 0; i--) {
             int nx = par[i][x];
             int ny = par[i][y];
             if (nx != ny) x = nx, y = ny;
